+++ extern/audaspace/src/util/RingBuffer.cpp
@@ -25,7 +25,7 @@
 
 AUD_NAMESPACE_BEGIN
 
-RingBuffer::RingBuffer(int size) :
+RingBuffer::RingBuffer(size_t size) :
 	m_buffer(size),
 	m_read(0),
 	m_write(0)
@@ -37,7 +37,7 @@ sample_t* RingBuffer::getBuffer() const
 	return m_buffer.getBuffer();
 }
 
-int RingBuffer::getSize() const
+size_t RingBuffer::getSize() const
 {
 	return m_buffer.getSize();
 }
@@ -122,13 +122,13 @@ void RingBuffer::reset()
 	m_write = 0;
 }
 
-void RingBuffer::resize(int size)
+void RingBuffer::resize(size_t size)
 {
 	m_buffer.resize(size);
 	reset();
 }
 
-void RingBuffer::assureSize(int size)
+void RingBuffer::assureSize(size_t size)
 {
 	m_buffer.assureSize(size);
 	reset();
